feat: add service to list users for admin
