+++ src/services/api.js
@@ -173,15 +173,4 @@ export const foldersAPI = {
   },
 };
 
-export const usersAPI = {
-  getAllUsers: async () => {
-    const response = await api.get("/users");
-    return response.data;
-  },
-  updateUser: async (userId, data) => {
-    const response = await api.patch(`/users/${userId}`, data);
-    return response.data;
-  },
-};
-
 export default api;
